Now deffering everything in a service injectable and usable wherever i want.
