+++ angular/src/app/login/login.component.html
@@ -8,11 +8,11 @@ register button :
 
 -->
 
-<div *ngIf="isLoggedIn()">
-    Logged in as {{user.username}}
+<div *ngIf="loginService.isLoggedIn()">
+    Logged in as {{loginService.user.username}}
     
 </div>
-<div *ngIf="!isLoggedIn()">
+<div *ngIf="!loginService.isLoggedIn()">
     <form id=loginForm (ngSubmit)="login()" #refForm="ngForm">
         <div>
             <input type="text" name="username" value="" placeholder="Username" [(ngModel)]="username" class="input">
